Change the type value passed to eomCallback, use index into parseTable

+++ Examples/Multiple_Parsers/Multiple_Parsers.ino
@@ -220,7 +220,7 @@ void loop()
 
 // Call back from within parser, for end of message
 // Process a complete message incoming from parser
-void nmeaMessage(SEMP_PARSE_STATE *parse, uint8_t type)
+void nmeaMessage(SEMP_PARSE_STATE *parse, uint16_t type)
 {
     SEMP_SCRATCH_PAD *scratchPad = (SEMP_SCRATCH_PAD *)parse->scratchPad;
     static bool displayOnce = true;
@@ -244,7 +244,7 @@ void nmeaMessage(SEMP_PARSE_STATE *parse, uint8_t type)
 
 // Call back from within parser, for end of message
 // Process a complete message incoming from parser
-void ubloxMessage(SEMP_PARSE_STATE *parse, uint8_t type)
+void ubloxMessage(SEMP_PARSE_STATE *parse, uint16_t type)
 {
     static bool displayOnce = true;
     uint32_t offset;
